Caching infra fixes
