+++ apps/api/data/src/infra/createDataApp.ts
@@ -1,110 +1,101 @@
-import * as path from 'path'
-import * as pulumi from '@pulumi/pulumi'
-import * as aws from '@pulumi/aws'
-
-import { createAppModule, PulumiApp, PulumiAppModule } from '@webiny/pulumi'
-import { CoreOutput } from '@webiny/pulumi-aws/apps/common'
-import { createLambdaRole, getCommonLambdaEnvVariables } from '@webiny/pulumi-aws/apps/lambdaUtils'
-import { getAwsAccountId, getAwsRegion } from '@webiny/pulumi-aws/apps/awsUtils'
 import { ApiGateway, ApiCloudfront } from '@webiny/pulumi-aws'
-import { DataApiGateway } from './createDataGateway'
 // import { DataApiCloudfront } from './createDataCloudfront'
-import { DataBucket } from './createDataBucket'
-import { createDataLambdaPolicy} from './createDataLambdaPolicy'
+import { type ApiPulumiApp } from '@webiny/pulumi-aws'
+import { CoreOutput } from '@webiny/pulumi-aws'
+import { DataFunction } from './createDataFunction'
+import { createDataBucket } from './createDataBucket'
+
 export interface DataApiParams {
   env: Record<string, any>
 }
-export type DataFunction = PulumiAppModule<ReturnType<typeof DataFunction>>
-export const DataFunction = (dataBucket: DataBucket) => {
-  return createAppModule({
-    name: 'PublicDataApi',
-    config(app: PulumiApp, params: DataApiParams) {
-      const dataResources = createDataResources(app, params, dataBucket)
-      return {
-        data: dataResources
-      }
+
+export const createDataApp = (app: ApiPulumiApp) => {
+  const { resources, getModule } = app
+
+  const { primaryDynamodbTableName, elasticsearchDomainEndpoint, elasticsearchDynamodbTableName } =
+    getModule(CoreOutput)
+  const name = 'api-data-bucket'
+
+  const { dataBucket, originIdentity } = createDataBucket(app, name)
+  const cloudfront = app.getModule(ApiCloudfront)
+
+  const {
+    data: {
+      dataApiGateway
+    }
+  } = app.addModule(DataFunction(dataBucket), {
+    env: {
+      DB_TABLE: primaryDynamodbTableName,
+      DB_TABLE_ELASTICSEARCH: elasticsearchDynamodbTableName,
+      ELASTIC_SEARCH_ENDPOINT: elasticsearchDomainEndpoint,
+      ELASTIC_SEARCH_INDEX_PREFIX: process.env.ELASTIC_SEARCH_INDEX_PREFIX,
+      S3_BUCKET: dataBucket.output,
+      API_URL: cloudfront.output.domainName
     }
   })
-}
-function createDataResources(app: PulumiApp, params: DataApiParams, dataBucket: DataBucket) {
-  if (!params.env['S3_BUCKET']) {
-    throw new Error(
-      'Missing S3_BUCKET environment variable. This is required for the data API to work.'
-    )
-  }
-
-  const core = app.getModule(CoreOutput)
-  const cacheTable = app.addResource(aws.dynamodb.Table, {
-    name: 'data-api-cache',
-    config: {
-      attributes: [
-        { name: 'PK', type: 'S' },
-        { name: 'SK', type: 'S' },
-        { name: 'fileid', type: 'S' },
-        { name: 'timestamp', type: 'N' }
-      ],
-      hashKey: 'PK',
-      rangeKey: 'SK',
-      billingMode: 'PAY_PER_REQUEST',
-      globalSecondaryIndexes: [
-        {
-          name: 'timestamp-index',
-          hashKey: 'fileid',
-          rangeKey: 'timestamp',
-          projectionType: 'ALL'
+  console.log('resources', resources.fileManager.functions.download.config.code)
+  resources.cloudfront.config.origins(origins => {
+    return [
+      ...origins,
+      {
+        domainName: dataApiGateway.stage.output.invokeUrl.apply(
+          (url: string) => new URL(url).hostname
+        ),
+        originPath: dataApiGateway.stage.output.invokeUrl.apply(
+          (url: string) => new URL(url).pathname
+        ),
+        originId: dataApiGateway.api.output.name,
+        customOriginConfig: {
+          httpPort: 80,
+          httpsPort: 443,
+          originProtocolPolicy: 'https-only',
+          originSslProtocols: ['TLSv1.2']
         }
-      ]
-    }
-  })
-
-  const policy = createDataLambdaPolicy(app, params, cacheTable.output.arn)
-  const role = createLambdaRole(app, {
-    name: 'data-api-lambda-role',
-    policy: policy.output
-  })
-  const dataQuery = app.addResource(aws.lambda.Function, {
-    name: 'data-api-runner',
-    config: {
-      role: role.output.arn,
-      runtime: 'nodejs14.x',
-      handler: 'handler.handler',
-      timeout: 60,
-      memorySize: 4096,
-      description: 'Runs data jobs for the Nectr data API',
-      code: new pulumi.asset.AssetArchive({
-        '.': new pulumi.asset.FileArchive(path.join(app.paths.workspace, 'data/build'))
-      }),
-      layers: ['arn:aws:lambda:us-east-2:217827289796:layer:duckdb-test:1'],
-      environment: {
-        variables: getCommonLambdaEnvVariables().apply(value => ({
-          ...value,
-          ...params.env,
-          S3_BUCKET: core.fileManagerBucketId,
-          EXTENSION_BUCKET: '',
-          DATA_BUCKET: pulumi.interpolate`${params.env['S3_BUCKET'].id}`,
-          COGNITO_USER_POOL_ID: core.cognitoUserPoolId,
-          CACHE_TABLE: cacheTable.output.name
-        }))
+      },
+      {
+        domainName: dataBucket.output.bucketRegionalDomainName,
+        originId: dataBucket.output.id,
+        s3OriginConfig: {
+          originAccessIdentity: originIdentity.output.cloudfrontAccessIdentityPath
+        }
       }
-    }
+    ]
   })
-
-  const dataApiGateway = app.addModule(DataApiGateway, {
-    'api-data-query': {
-      path: '/data-query/{id}',
-      method: 'ANY',
-      function: dataQuery.output.arn
+  resources.cloudfront.config.orderedCacheBehaviors(orderedBehaviors => {
+    if (!orderedBehaviors) {
+      return orderedBehaviors
     }
+    const behaviors = orderedBehaviors.map(behavior => behavior)
+    behaviors.push({
+      compress: true,
+      allowedMethods: ['GET', 'HEAD', 'OPTIONS', 'PUT', 'POST', 'PATCH', 'DELETE'],
+      cachedMethods: ['GET', 'HEAD', 'OPTIONS'],
+      forwardedValues: {
+        cookies: {
+          forward: 'none'
+        },
+        headers: ['Accept', 'Accept-Language'],
+        queryString: true
+      },
+      pathPattern: '/data-query*',
+      viewerProtocolPolicy: 'allow-all',
+      targetOriginId: dataApiGateway.api.output.name.apply(v => `${v}`) as unknown as string
+    })
+    behaviors.push({
+      compress: true,
+      allowedMethods: ['GET', 'HEAD', 'OPTIONS'],
+      cachedMethods: ['GET', 'HEAD', 'OPTIONS'],
+      forwardedValues: {
+        cookies: {
+          forward: 'none'
+        },
+        headers: ['Accept', 'Accept-Language'],
+        queryString: false
+      },
+      pathPattern: '/cache*',
+      viewerProtocolPolicy: 'allow-all',
+      targetOriginId: dataBucket.output.id.apply(v => `${v}`) as unknown as string
+    })
+    return behaviors
   })
-  
-  return {
-    role,
-    policy,
-    cacheTable,
-    dataApiGateway,
-    functions: {
-      dataQuery
-    }
-  }
 }
-
